+++ src/main/java/school/hei/haapi/model/dto/MissedEventStatsDto.java
@@ -2,11 +2,11 @@ package school.hei.haapi.model.dto;
 
 import school.hei.haapi.endpoint.rest.model.MissedEventStats;
 
-public record MissedEventStatsDto(Long total, Long justifiedAbsence, Long unjustifiedAbsence) {
+public record MissedEventStatsDto(long total, long justifiedAbsence, long unjustifiedAbsence) {
   public MissedEventStats toRest() {
     return new MissedEventStats()
-        .justified(this.justifiedAbsence.intValue())
-        .unjustified(this.unjustifiedAbsence.intValue())
-        .total(this.total.intValue());
+        .justified(this.justifiedAbsence)
+        .unjustified(this.unjustifiedAbsence)
+        .total(this.total);
   }
 }
